+++ src/autogain.cpp
@@ -34,6 +34,18 @@ AutoGain::AutoGain(const std::string& tag,
     init_ebur128();
   });
 
+  settings->signal_changed("reference").connect([&, this](const auto& key) {
+    if (settings->get_string(key) == "Momentary") {
+      reference = Reference::momentary;
+    } else if (settings->get_string(key) == "Shortterm") {
+      reference = Reference::shortterm;
+    } else if (settings->get_string(key) == "Integrated") {
+      reference = Reference::integrated;
+    } else if (settings->get_string(key) == "Geometric Mean") {
+      reference = Reference::geometric_mean;
+    }
+  });
+
   setup_input_output_gain();
 }
 
@@ -150,7 +162,28 @@ void AutoGain::process(std::span<float>& left_in,
     }
 
     if (!failed) {
-      loudness = std::cbrt(momentary * shortterm * global);
+      switch (reference) {
+        case Reference::momentary: {
+          loudness = momentary;
+
+          break;
+        }
+        case Reference::shortterm: {
+          loudness = shortterm;
+
+          break;
+        }
+        case Reference::integrated: {
+          loudness = global;
+
+          break;
+        }
+        case Reference::geometric_mean: {
+          loudness = std::cbrt(momentary * shortterm * global);
+
+          break;
+        }
+      }
 
       const double diff = target - loudness;
 
@@ -159,7 +192,7 @@ void AutoGain::process(std::span<float>& left_in,
 
       const double peak = (peak_L > peak_R) ? peak_L : peak_R;
 
-      const auto& db_peak = util::linear_to_db(peak);
+      const auto db_peak = util::linear_to_db(peak);
 
       if (db_peak > util::minimum_db_level) {
         if (gain * peak < 1.0) {