+++ src/equalizer_ui.cpp
@@ -596,6 +596,22 @@ void EqualizerUi::on_import_apo_preset_clicked() {
 }
 
 // returns false if we cannot parse given line successfully
+bool EqualizerUi::parse_apo_preamp(const std::string& line, double& preamp) {
+  std::smatch matches;
+
+  static const auto i = std::regex::icase;
+
+  std::regex_search(line, matches, std::regex(R"(preamp:\s*+([+-]?+\d++\.?+\d*+)\s*+db)", i));
+
+  if (matches.size() != 2U) {
+    return false;
+  }
+
+  preamp = std::stod(matches.str(1));
+  return true;
+}
+
+// returns false if we cannot parse given line successfully
 
 auto EqualizerUi::parse_apo_filter(const std::string& line, struct ImportedBand& filter) -> bool {
   std::smatch matches;
@@ -630,7 +646,7 @@ auto EqualizerUi::parse_apo_filter(const std::string& line, struct ImportedBand&
 
   if ((filter.type & (LOW_SHELF_xdB | HIGH_SHELF_xdB | LOW_SHELF | HIGH_SHELF)) != 0U) {
     std::regex_search(line, matches,
-                      std::regex(R"(filter\s++\d*+:\s*+on\s++[a-z]++\s++([\+-]?+\d++\.?+\d*+)\s*+db)", i));
+                      std::regex(R"(filter\s++\d*+:\s*+on\s++[a-z]++\s++([+-]?+\d++\.?+\d*+)\s*+db)", i));
 
     // _xdB variants require the dB parameter
 
@@ -648,7 +664,7 @@ auto EqualizerUi::parse_apo_filter(const std::string& line, struct ImportedBand&
   // get gain
 
   if ((filter.type & (PEAKING | LOW_SHELF_xdB | HIGH_SHELF_xdB | LOW_SHELF | HIGH_SHELF)) != 0U) {
-    std::regex_search(line, matches, std::regex(R"(gain\s++([\+-]?+\d++\.?+\d*+)\s*+db)", i));
+    std::regex_search(line, matches, std::regex(R"(gain\s++([+-]?+\d++\.?+\d*+)\s*+db)", i));
 
     // all Shelf types (i.e. all above except for Peaking) require the gain parameter
 
@@ -685,6 +701,7 @@ void EqualizerUi::import_apo_preset(const std::string& file_path) {
   if (std::filesystem::is_regular_file(p)) {
     std::ifstream eq_file;
     std::vector<struct ImportedBand> bands;
+    double preamp = 0.0;
 
     eq_file.open(p.c_str());
 
@@ -694,8 +711,10 @@ void EqualizerUi::import_apo_preset(const std::string& file_path) {
       while (getline(eq_file, line)) {
         struct ImportedBand filter {};
 
-        if (this->parse_apo_filter(line, filter)) {
-          bands.push_back(filter);
+        if (!parse_apo_preamp(line, preamp)) {
+          if (parse_apo_filter(line, filter)) {
+            bands.push_back(filter);
+          }
         }
       }
     }
@@ -707,6 +726,7 @@ void EqualizerUi::import_apo_preset(const std::string& file_path) {
     }
 
     settings->set_int("num-bands", bands.size());
+    settings->set_double("input-gain", preamp);
 
     for (int n = 0; n < max_bands; n++) {
       const auto& bandn = "band" + std::to_string(n);